backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
supabase==2.10.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT[crypto]==2.10.1
